updated status code of some situations with the tests

+++ routes/delCatDetails.js
@@ -1,21 +1,39 @@
 var express = require("express");
 var router = express.Router();
 
-var express = require("express");
-var router = express.Router();
-
 // Import the db object
 var db = require("../dbConnect");
 
 function deleteCatDetails(catID, callback) {
+  // First check if the cat with the given catID exists
   db.query(
-    "DELETE FROM Cats WHERE catID = ?",
+    "SELECT * FROM Cats WHERE catID = ?",
     [catID],
     function (error, results, fields) {
       if (error) {
-        callback({ status: 500, error: error, response: null });
+        callback({ status: 500, error: error });
       } else {
-        callback({ status: 200, error: null, response: 'Cat details has been removed' });
+        // If no records found, send a 404 response
+        if (results.length === 0) {
+          callback({ status: 404, error: "Cat not found" });
+        } else {
+          // If the record exists, attempt to delete it
+          db.query(
+            "DELETE FROM Cats WHERE catID = ?",
+            [catID],
+            function (error, results, fields) {
+              if (error) {
+                callback({ status: 500, error: error });
+              } else {
+                // Send a 200 response indicating the record was deleted
+                callback({
+                  status: 200,
+                  response: "Cat details have been removed",
+                });
+              }
+            }
+          );
+        }
       }
     }
   );
@@ -24,7 +42,7 @@ function deleteCatDetails(catID, callback) {
 router.delete("/:catID", function (req, res, next) {
   const catID = req.params.catID;
   deleteCatDetails(catID, function (response) {
-    res.send(response);
+    res.status(response.status).send(response);
   });
 });
